Auto scale up if execution time is very small.
Example: if a function execution time is really small (under 0.01s), it will
try to use a bigger scale (multiplying the number of calls by a factor of 10)
and re-run the function until it executes in more than 0.01s.

This procedure will be executed for as many as 5 times, possibly running the
function for a maximum of 100_000 times. If it fails to find an execution time
bigger than 0.01s using this method, it will throw an exception.