Remove everything that is not 100% stable.
SpaceComplexity is one of them. It was particularly unstable with 1.8.7, and there
was no obvious way of fixing it.

SpaceComplexity used a fork and `ps` to measure the memory usage. We can't use a
profiler since one of the requirement is to patch the Ruby Interpreter (and I would
avoid that as much as possible). Furthermore, it is not possible to get the memory
footprint of an object.